+++ src/utils/types.rs
@@ -1,5 +1,5 @@
+use crate::network::client::Client;
 use std::{collections::HashMap, sync::Arc};
 use tokio::sync::RwLock;
-use crate::network::client::Client;
 
-pub type ClientPool = Arc<RwLock<HashMap<u64, Arc<Client>>>>;
+pub type ClientPool = Arc<RwLock<HashMap<i32, Arc<Client>>>>;
